Add support for IO::Buffer in #send

+++ ext/um/um.h
@@ -250,7 +250,7 @@ VALUE um_accept(struct um *machine, int fd);
 VALUE um_accept_each(struct um *machine, int fd);
 VALUE um_socket(struct um *machine, int domain, int type, int protocol, uint flags);
 VALUE um_connect(struct um *machine, int fd, const struct sockaddr *addr, socklen_t addrlen);
-VALUE um_send(struct um *machine, int fd, VALUE buffer, int len, int flags);
+VALUE um_send(struct um *machine, int fd, VALUE buffer, size_t len, int flags);
 VALUE um_send_bundle(struct um *machine, int fd, int bgid, VALUE strings);
 VALUE um_recv(struct um *machine, int fd, VALUE buffer, int maxlen, int flags);
 VALUE um_recv_each(struct um *machine, int fd, int bgid, int flags);
